+++ packages/flame/CHANGELOG.md
@@ -28,7 +28,7 @@
  - Renamed `FlameTester` to `GameTester`
  - Modified `FlameTester` to be specific for `T extends FlameGame`
  - Improved `TimerComponent`
- - Method `preRender()` and `postRender()` are now deprecated
+ - Removed methods `preRender()` and `postRender()` in `Component`
 
 ## [1.0.0-releasecandidate.16]
  - `changePriority` no longer breaks game loop iteration